requirements.txt
rapidfuzz
python3-apt